useless correction

+++ CFactorialCalculation.cs
@@ -8,9 +8,11 @@ namespace OmSULab
 {
     internal class CFactorialCalculation
     {
-        public static int iFactorialCalc(int n)
+        public static double iFactorialCalc(double n)
         {
-            if (n == 0 || n == 1) return 1;
+            if (n == 0 || n == 1) 
+                return 1;
+
             return n * iFactorialCalc(n - 1);
         }
     }
